+++ src/lib/types/reservation/create-reservation.type.ts
@@ -5,3 +5,11 @@ export interface CreateReservationType {
     phone: string;
     quantity: number;
 }
+
+export interface CreateReservationTemporalType {
+    date?: string;
+    time?: string;
+    name: string;
+    phone: string;
+    quantity: string;
+}